Update a library function definition to include a "hash_algo" parameter
Instead of relying on a global "hash_algo" variable, which is not
concurrency-safe, update the verify_hash2() function definition to
include a "hash_algo" parameter as a place holder.

Export the verify_hash2() definition.

Define verify_hash2().  To avoid library incompatablity, make the existing
function verify_hash() a function wrapper for verify_hash2().

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>