Define library ima_calc_hash2() function with a hash algorithm parameter
Instead of relying on the "imaevm_params.algo" global variable, which
is not concurrency-safe, define a new library ima_calc_hash2() function
with the hash algorithm as a parameter.

To avoid library incompatablity, make the existing ima_calc_hash()
function a wrapper for ima_calc_hash2().

Deprecate ima_calc_hash().

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>