Pass ENGINE and keyid through to function using them
Pass the ENGINE and keyid all the way through to the function that is
using them and deprecate sign_hash since it needs to pass these parameters
from the global imaevm_params.

In anticipation of support for OpenSSL providers, wrap the ENGINE in a
union inside a struct imaevm_ossl_access and add a type file dfor the
selection of the ENGINE or provider later on.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>